requirements.txt
streamlit==1.31.1
mysql-connector-python==8.3.0
Pillow==10.2.0
python-dotenv==1.0.1
razorpay==1.4.1
streamlit-autorefresh==1.0.1 